Moved chatTheme convenience initializers into thier own file.
